moved common class to commons module
