+++ gazeplay-commons/src/main/java/net/gazeplay/commons/utils/games/ImageDirectoryLocator.java
@@ -1,4 +1,4 @@
-package net.gazeplay.games.whereisit;
+package net.gazeplay.commons.utils.games;
 
 import lombok.extern.slf4j.Slf4j;
